improve

+++ fddn-contact/fddn_contact/views/contact_update.py
@@ -7,15 +7,27 @@ from ..models import Contact, ContactInfoItem
 def contact_update_view(request):
     contact_data = request.json_body
 
-    request.dbsession\
-        .query(ContactInfoItem)\
-        .filter(ContactInfoItem.contact_id == contact_data['id'])\
-        .delete()
+    if 'id' in contact_data:
+        contact = request.dbsession \
+            .query(Contact) \
+            .filter(Contact.id == contact_data['id']) \
+            .first()
+
+        if contact is None:
+            return {'result': 'failed', 'reason': 'object not found'}
+
+        request.dbsession\
+            .query(ContactInfoItem)\
+            .filter(ContactInfoItem.contact_id == contact_data['id'])\
+            .delete()
+    else:
+        pass
 
     request.dbsession\
         .query(Contact)\
         .filter(Contact.id == contact_data['id'])\
         .update({Contact.name: contact_data['name']})
+
     for i, item_data in enumerate(contact_data['items']):
         item = ContactInfoItem()
         item.contact_id = contact_data['id']
@@ -24,4 +36,4 @@ def contact_update_view(request):
         item.index = i
         request.dbsession.add(item)
 
-    return {"result":"OK"}
+    return {"result": "ok"}
